requirements.txt
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
Django==5.1.6
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
djangorestframework==3.15.2
einops==0.8.1
entmax==1.3
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
munch==4.0.0
mysqlclient==2.2.7
networkx==3.4.2
numpy
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pix2tex==0.1.4
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
response==0.5.0
safetensors==0.5.3
scipy==1.15.2
setuptools==75.8.2
simsimd==6.2.1
six==1.17.0
sqlparse==0.5.3
stringzilla==3.12.1
sympy==1.13.1
timm==0.5.4
tokenizers==0.21.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
x-transformers==0.15.0
